Added the dashboards for physicians and staff. Added a service request details screen. When the user logs in, they are directed to their dashboards.

+++ WebApp/Controllers/ServiceRequestController.cs
@@ -22,7 +22,7 @@ namespace WebApp.Controllers
         }
 
         // GET: ServiceRequest/Details/5
-        public async Task<ActionResult> Details(int? id)
+        public async Task<ActionResult> Details(short? id)
         {
             if (id == null)
             {
